Support routers that only provide WANIPService:2
InternetGatewayDevice:2 adds WANConnectionDevice:2 as a sub-device
that implements WANIPConnection:2 instead of WANIPConnection:1.

This provides the same methods as WANIPConnection:1 but has some
additional constraints such as the gateway must only serve it on the
LAN interface and should reject mappings to an address that wasn't the
origin of the request.

Signed-off-by: Richard Maw <richard.maw@gmail.com>